:sparkles: Update time-related structures and main function comments
- Changed `Hour` and `Timezone` fields in `RequestClockingBody` and `ResponseGetTime` structs from `int16` and `int8` to `uint32` and `uint8`, respectively, for better data representation.
- Updated comments in `main.go` to reflect changes in time duration calculations and added new date formatting logic.
- Improved clarity of commented-out code for future reference and potential reactivation.

+++ rhApi/structs.go
@@ -13,14 +13,14 @@ type IsLoggedResponse struct {
 
 type RequestClockingBody struct {
 	Date string `json:"date"`
-	Hour int16 `json:"hour"`
+	Hour uint32 `json:"hour"`
 	Latitude string `json:"latitude"`
 	Longitude string `json:"longitude"`
-	Timezone int8 `json:"timezone"`
+	Timezone uint8 `json:"timezone"`
 	Address string `json:"address"`
 }
 
 type ResponseGetTime struct {
 	ActualDate string `json:"actualDate"`
-	ActualTime int16 `json:"actualTime"`
+	ActualTime uint32 `json:"actualTime"`
 }
